refactor: integrate root headers
